Software Manager: refactor splitting into smaller modules
This should bring no change in behavior, just a finer separation
of the code.
